requirements.txt
python-dotenv
requests
matplotlib
pillow
numpy
